Added basic nav structure
Hey Brigid, about half of these links are active so far and navigate to
outside pages, not sure how many sub pages we’ll build so and dead
links we can leave unclickable or link back to the index if we want. I
also started a footer tag at the bottom of the body.