MEGA UPDATE
Hey all, I tried getting busy with a drop down nav with some mildly
tricky transitions. Turned out to be a fools errand, but I kept some of
what I worked on. Basically it was unresponsive, but I learned a couple
things heh. I commented out the drop down styles you want to look at
that mess, they are heavily edited.

I currently only have the top level navigation active, all of the sub
level nav is set to display:none; currently. The nav looks pretty good
until the 768px breakpoint, so we’ll come up with a good responsive
navigation strategy for the smaller breakpoint.

I established media queries for 768px and below, 769px to 1200px, and
1201px and up.

I started setting up the footer, obviously needs a lot of work. The
SVGs are in there, just need to tweak em and discuss what else we want
there as we moved the “about” and “contact” to the main nav.

I also employed the responsive logo which is looking pretty sharp to me.

I added an expanding animated search field in the header. Set a break
point to change the width percentage at the 768px media query. I need
to figure out a way to shave a bit off the right margin, though there
is no right margin so I’m not sure what’s pushing it away from the left
edge. No worries though.

I indented the html a bit to make it more readable, when doing this I
looked at the way we have columns structured, I think we may need to
revise the div/column organization a bit. Also do we have anymore
content wrangled up currently? If the intention was to go with
“progressive enhancement” model there were a few things that weren’t
gonna work for mobile style wise, we can go over it on Monday. So
currently we have some larger screen and smaller screen attributes
going on at the moment, which isn’t bad, just something to be conscious
of.

One more thing to consider, if we want to have a carousel/image slider
for mobile we may need to employ some javascript. I’m game just a heads
up.

All in all I think we’re still in pretty good shape and right on track.
I’ll shoot an email tomorrow, I think I’m gonna focus on the footer
style next and leave the nav as is for a bit.