Footer style and
Hey Brigid, I styled the footer a bit, the SVGs are clinging to the
left edge of the link they’re contained in so I’m gonna ask Erik about
that tomorrow, but I think they’re good enough to show. Can’t get a
hover state going on em yet, but we can figure that out later. Got the
grey contain to go around all the locations as well. Put it locations
into two columns just to see how it would work.