Fixed a couple JS bugs
I enabled hoverintent for the nav. this makes hover animation not fire
quite immediately, once the cursor speed slows it fires, so that way if
someone is going to click on the logo or search field the menu won’t
fly out and get in  the way or distract them.

I also commented out the jquery library calls you put in for the
carousel, we are already calling a library earlier on and I believe the
later library calls were neutralizing the initializations and calls
earlier.